Reviewing PCB and updating switch spacing
Changed soldermask layer on the connector to not be along the whole
trace. Reviewed the pinout of all the switches and confirmed they're
good. Reviewed the spacing of the switches and found errors, I have
moved them but am yet to reroute the traces. Updated the teardrops to be
curved.